+++ resources/views/documents/projects/layout.blade.php
@@ -281,6 +281,8 @@
 				@if(isset($facets)) facets: {!!json_encode($facets)!!}, @endif
 				@if(isset($filters)) filters: {!!json_encode($filters)!!}, @endif
 				isSearchRequest: {{ isset($is_search_requested) && $is_search_requested ? 'true' : 'false' }},
+				canShare: {{ isset($can_share) && $can_share ? 'true' : 'false' }},
+				canPublish: {{ isset($can_make_public) && isset($is_klink_public_enabled) && $is_klink_public_enabled && $can_make_public ? 'true' : 'false' }},
 				userIsProjectManager: {{ auth()->check() && auth()->user()->isProjectManager() ? 'true' : 'false' }}
 			});
 			Documents.groups.ensureCurrentVisibility();